requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
python-dotenv==1.0.1
httpx==0.27.0
pytest==7.4.4
moviepy==1.0.3
openai==1.12.0
python-multipart==0.0.9
python-telegram-bot==21.11.1
sqlite3-api==0.1.0
aiosqlite==0.19.0
tqdm==4.66.2 